Removed no more used i18n.Logger \o/ \o/
